# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"APIResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {},
						"type": "array"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"error": {
						"items": {},
						"type": "array"
					}
				},
				"required": [
					"message",
					"data",
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ShekiAI API",
		"version": "1.0.0",
		"description": "ShekiAI Backend API Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support",
			"email": "support@shekia.com"
		}
	},
	"paths": {
		"/auth/health": {
			"get": {
				"operationId": "Health",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ai_v1/course-draft/start": {
			"post": {
				"operationId": "Start",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Course Draft AI"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"tutorName": {
										"type": "string"
									},
									"tutorId": {
										"type": "string"
									}
								},
								"required": [
									"tutorName",
									"tutorId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ai_v1/course-draft/{sessionId}/message": {
			"post": {
				"operationId": "Message",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Course Draft AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"tutorName": {
										"type": "string"
									},
									"tutorId": {
										"type": "string"
									}
								},
								"required": [
									"message",
									"tutorName",
									"tutorId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ai_v1/course-draft/{sessionId}": {
			"get": {
				"operationId": "GetSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Course Draft AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tutorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ai_v1/course-draft/mine/list": {
			"get": {
				"operationId": "ListMine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Course Draft AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tutorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/ai_v1/course-draft/{sessionId}/abandon": {
			"post": {
				"operationId": "Abandon",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Course Draft AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tutorId": {
										"type": "string"
									}
								},
								"required": [
									"tutorId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ai_v1/ai-health": {
			"get": {
				"operationId": "Health",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"AI Controller"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ai_v1/test-prompt": {
			"post": {
				"operationId": "TestPrompt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AI Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"prompt": {
										"type": "string"
									}
								},
								"required": [
									"prompt"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ai_v1/translate-language": {
			"post": {
				"operationId": "TranslateLanguage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AI Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"languageCode": {
										"type": "string"
									},
									"lang": {
										"type": "string"
									},
									"text": {
										"type": "string"
									}
								},
								"required": [
									"languageCode",
									"lang",
									"text"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/ai_v1/ai-fetch-users": {
			"get": {
				"operationId": "FetchUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AI Controller"
				],
				"security": [],
				"parameters": []
			}
		},
		"/ai_v1/search": {
			"post": {
				"operationId": "SearchAgentController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"AI Controller"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"query": {
										"type": "string"
									}
								},
								"required": [
									"query"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}